--- SugarRushSite/src/app/meal-info/meal-info.component.html
@@ -1,11 +1,0 @@
-<div>
-   <div class="mealTitle">
-      {{ item.meal }}
-   </div>
-   <div class="foodItem">
-      {{ item.selectedFood.name }}
-   </div>
-   <div class="sugarAmount">
-      {{ item.selectedFood.sugarAmount }} grams
-   </div>
-</div>
