+++ SugarRushSite/src/app/meal-modal/meal-info/meal-info.component.html
@@ -1,0 +1,14 @@
+<div>
+   <div class="mealImage">
+      <img src="./assets/{{mealFood.imagePath}}.png">
+   </div>
+   <div class="mealTitle">
+      {{ mealTitle }}
+   </div>
+   <div class="foodItem">
+      {{ mealFood.name }}
+   </div>
+   <div class="sugarAmount">
+      {{ mealFood.sugarAmount }} grams
+   </div>
+</div>
